feat(book.ts): add default values for _id and comments properties in the book store
The changes were made to the `book.ts` file in the `store` directory of the `frontend` app. The default values for the `_id` and `comments` properties were added to the `book` store. This ensures that when a new book is created, it will have a default value of 0 for `_id` and an empty array for `comments`.