feat(users): add user profile page component
This commit adds a new Svelte component for the user profile page. The component receives a `data` prop which contains information about the user. If the user is found in the data, the component displays the user's avatar, username, headline, and about section. If the user is not found, a "User not found" message is displayed.