feat(users/[slug]/+page.svelte): add functionality to handle following a user and display follow count
The changes in this commit add functionality to handle following a user in the users/[slug]/+page.svelte file. The getCurrentUserId function is added to get the current user's ID. The handleFollow function is added to handle the follow action when the "Follow" button is clicked. The follow count is also displayed in the UI using the data.followeesCount and data.followersCount variables.